Checkpoint - isize -> i32

+++ alacritty/src/display/content.rs
@@ -128,7 +128,7 @@ impl<'a> RenderableContent<'a> {
 
         // Convert cursor point to viewport position.
         let mut point = self.terminal_cursor.point;
-        point.line += self.terminal_content.display_offset as isize;
+        point.line += self.terminal_content.display_offset as i32;
 
         Some(RenderableCursor {
             shape: self.terminal_cursor.shape,
@@ -244,7 +244,7 @@ impl RenderableCell {
 
         // Convert cell point to viewport position.
         let mut point = cell.point;
-        point.line += content.terminal_content.display_offset as isize;
+        point.line += content.terminal_content.display_offset as i32;
 
         RenderableCell {
             zerowidth: cell.zerowidth().map(|zerowidth| zerowidth.to_vec()),
@@ -432,8 +432,8 @@ pub struct RegexMatches(Vec<RangeInclusive<Point>>);
 impl RegexMatches {
     /// Find all visible matches.
     pub fn new<T>(term: &Term<T>, dfas: &RegexSearch) -> Self {
-        let viewport_start = Line(-(term.grid().display_offset() as isize));
-        let viewport_end = viewport_start + term.screen_lines() - 1isize;
+        let viewport_start = Line(-(term.grid().display_offset() as i32));
+        let viewport_end = viewport_start + term.screen_lines() - 1;
 
         // Compute start of the first and end of the last line.
         let start_point = Point::new(viewport_start, Column(0));
